cleared main.py, all the api code is now in api directory and all the func in myfunc.py
