+++ src/back/apiRequests/verify_reset_code.py
@@ -1,13 +1,20 @@
-from flask import request, jsonify
+import enc
+from flask import Blueprint, request, jsonify
+from models import OTP
 
-reset_codes = {}  # This should be a shared state if used in multiple modules
+verify_reset_code_bp = Blueprint('verify_reset_code_bp', __name__)
 
+@verify_reset_code_bp.route('/verify_reset_code', methods=['POST'])
 def verify_reset_code():
     data = request.get_json()
     email = data.get('email')
     code = data.get('code')
 
-    if reset_codes.get(email) == code:
-        return jsonify({'message': 'Code verified', 'status': 200}), 200
+    user_otp = OTP.query.filter_by(email=email).first()
+    if user_otp is None:
+        return jsonify({'message': 'User not found', 'status': 404})
+    secret_key = user_otp.secret_key
+    if enc.verify_otp(secret_key, code):
+        return jsonify({'message': 'Code verified', 'status': 200})
     else:
-        return jsonify({'message': 'Invalid code or expired', 'status': 400}), 400
+        return jsonify({'message': 'Invalid code or expired', 'status': 400})
